graphpop-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graphpop-cli"
version = "0.1.0"
description = "Command-line interface for GraphPop population genomics procedures"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "neo4j>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
graphpop = "graphpop_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
